refactor(qgrads): refactored interface for autograd and mathoperations
