+++ elasticai/creator/nn/quantized_grads/base_modules/sigmoid.py
@@ -1,7 +1,7 @@
 import torch
 from torch.nn import Sigmoid as _Sigmoid
 
-from elasticai.creator.nn.quantized_grads._math_operations import MathOperations
+from elasticai.creator.nn.quantized_grads.math_operations import MathOperations
 
 
 class Sigmoid(_Sigmoid):
